moved timer to eventmanager
